[FUN-3507] Wired up real padre to the content optimiser, enabling arbitrary queries to search/content-optimiser.html. The code is now in the SearchController, and will ultimately need to be moved into a plugin instead.
git-svn-id: https://svn.cbr.au.funnelback.com/svn/enterprise-search/trunk/java/publicui@23914 3cdb6afb-a121-0410-a22c-8da0070c90a1

+++ publicui-web/src/main/java/com/funnelback/contentoptimiser/RankingScore.java
@@ -6,5 +6,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class RankingScore {
 	@Getter private final String name;	
-	@Getter private final int percentage;
+	@Getter private final float percentage;
 }
